Move DisjointSet to file
